Create react app to vite
