+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@vitejs/plugin-react": "^4.3.0",
     "ol": "^7.4.0",
     "proj4": "^2.9.0",
     "react": "^18.2.0",
@@ -10,12 +11,13 @@
     "react-responsive-carousel": "^3.2.23",
     "react-router": "^6.14.2",
     "react-router-dom": "^6.14.2",
+    "vite": "^5.2.12",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
+    "start": "vite --host 0.0.0.0",
+    "build": "vite build",
+    "preview": "vite preview",
     "eject": "react-scripts eject",
     "lint:js": "eslint --ext .js --ext .jsx ./src",
     "lint:js:fix": "eslint --ext .js --ext .jsx --fix ./src",
@@ -42,6 +44,7 @@
   "devDependencies": {
     "@babel/eslint-parser": "^7.22.9",
     "@tailwindcss/line-clamp": "^0.4.4",
+    "autoprefixer": "^10.4.19",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-cypress": "^2.13.3",
@@ -53,7 +56,6 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "postcss": "^8.4.27",
     "prettier": "^2.8.8",
-    "react-scripts": "5.0.1",
     "tailwindcss": "^3.3.3"
   }
 }
