+++ src/components/tags/TagsList.jsx
@@ -1,7 +1,7 @@
 import { React, useEffect, useState } from "react";
 import useFetch from "../../util/useFetch";
 import Tag from "./Tag";
-import { ReactComponent as Xmark } from "../../icons/xmark-solid.svg";
+import Xmark from "../../icons/xmark-solid.svg";
 
 function TagsList() {
   const [tags, setTags] = useState(null);
@@ -10,6 +10,7 @@ function TagsList() {
 
   useEffect(() => {
     if (data) {
+      console.log(data);
       setTags(data["hydra:member"]);
     }
   }, [data]);
