--- src/index.js
@@ -1,32 +1,0 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { BrowserRouter } from "react-router-dom";
-import ApiEndpointContext from "./context/api-endpoint-context";
-import reportWebVitals from "./reportWebVitals";
-import App from "./App";
-import "./index.css";
-
-const root = ReactDOM.createRoot(document.getElementById("root"));
-
-root.render(
-  <React.StrictMode>
-    <BrowserRouter>
-      <ApiEndpointContext.Provider
-        value={{
-          url: `${process.env.REACT_APP_API_BASE}${process.env.REACT_APP_API_ROUTE}`,
-          fileUrl: process.env.REACT_APP_API_BASE,
-          token: process.env.REACT_APP_TOKEN,
-          mapUsername: process.env.REACT_APP_DF_MAP_USERNAME,
-          mapPassword: process.env.REACT_APP_DF_MAP_PASSWORD,
-        }}
-      >
-        <App />
-      </ApiEndpointContext.Provider>
-    </BrowserRouter>
-  </React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
